refactor: restructure LandScheduler with pimpl pattern
