--- src/pland/land/LandScheduler.cc
@@ -1,235 +1,0 @@
-#include "pland/land/LandScheduler.h"
-#include "ll/api/chrono/GameChrono.h"
-#include "ll/api/coro/CoroTask.h"
-#include "ll/api/event/EventBus.h"
-#include "ll/api/event/ListenerBase.h"
-#include "ll/api/event/player/PlayerDisconnectEvent.h"
-#include "ll/api/event/player/PlayerJoinEvent.h"
-#include "ll/api/service/Bedrock.h"
-#include "ll/api/service/PlayerInfo.h"
-#include "ll/api/thread/ServerThreadExecutor.h"
-#include "mc/network/packet/SetTitlePacket.h"
-#include "mc/server/ServerPlayer.h"
-#include "mc/world/actor/player/Player.h"
-#include "mc/world/level/ChunkPos.h"
-#include "mc/world/level/Level.h"
-#include "pland/Global.h"
-#include "pland/PLand.h"
-#include "pland/events/player/PlayerMoveEvent.h"
-#include "pland/infra/Config.h"
-#include "pland/land/LandEvent.h"
-#include "pland/land/repo/LandRegistry.h"
-#include <cstdio>
-#include <stdexcept>
-#include <vector>
-
-
-namespace land {
-
-
-LandScheduler::LandScheduler() {
-    auto& bus = ll::event::EventBus::getInstance();
-
-    mQuit                   = std::make_shared<std::atomic<bool>>(false);
-    mEventSchedulingSleep   = std::make_shared<ll::coro::InterruptableSleep>();
-    mLandTipSchedulingSleep = std::make_shared<ll::coro::InterruptableSleep>();
-
-    mPlayerJoinServerListener = bus.emplaceListener<ll::event::PlayerJoinEvent>([this](ll::event::PlayerJoinEvent& ev) {
-        auto& player = ev.self();
-        if (player.isSimulatedPlayer()) {
-            return;
-        }
-        mPlayers.emplace_back(&player);
-    });
-
-    mPlayerDisconnectListener =
-        bus.emplaceListener<ll::event::PlayerDisconnectEvent>([this](ll::event::PlayerDisconnectEvent& ev) {
-            auto& player = ev.self();
-            if (player.isSimulatedPlayer()) {
-                return;
-            }
-
-            auto ptr = &player;
-            mDimensionMap.erase(ptr);
-            mLandIdMap.erase(ptr);
-            std::erase_if(mPlayers, [&ptr](auto* p) { return p == ptr; });
-        });
-
-    mPlayerEnterLandListener = bus.emplaceListener<event::PlayerEnterLandEvent>([](event::PlayerEnterLandEvent& ev) {
-        if (!Config::cfg.land.tip.enterTip) {
-            return;
-        }
-
-        auto& player   = ev.self();
-        auto& registry = PLand::getInstance().getLandRegistry();
-
-        if (auto settings = registry.getPlayerSettings(player.getUuid()); settings && !settings->showEnterLandTitle) {
-            return; // 如果玩家设置不显示进入领地提示,则不显示
-        }
-
-        auto land = registry.getLand(ev.landId());
-        if (!land) {
-            return;
-        }
-
-        SetTitlePacket title(SetTitlePacket::TitleType::Title);
-        SetTitlePacket subTitle(SetTitlePacket::TitleType::Subtitle);
-
-        if (land->isOwner(player.getUuid())) {
-            title.mTitleText    = land->getName();
-            subTitle.mTitleText = "欢迎回来"_trf(player);
-        } else {
-            title.mTitleText    = "Welcome to"_trf(player);
-            subTitle.mTitleText = land->getName();
-        }
-
-        title.sendTo(player);
-        subTitle.sendTo(player);
-    });
-
-    ll::coro::keepThis([quit = mQuit, sleep = mEventSchedulingSleep, this]() -> ll::coro::CoroTask<> {
-        while (!quit->load()) {
-            co_await sleep->sleepFor(ll::chrono::ticks{5});
-            if (quit->load()) {
-                break;
-            }
-
-            if (mPlayers.empty()) {
-                continue;
-            }
-
-            try {
-                tickEvent();
-            } catch (std::exception& e) {
-                PLand::getInstance().getSelf().getLogger().error(
-                    "An exception occurred while scheduling land events: {}",
-                    e.what()
-                );
-            } catch (...) {
-                PLand::getInstance().getSelf().getLogger().error(
-                    "An unknown exception occurred while scheduling land events."
-                );
-            }
-        }
-        co_return;
-    }).launch(ll::thread::ServerThreadExecutor::getDefault());
-
-    if (Config::cfg.land.tip.bottomContinuedTip) {
-        ll::coro::keepThis([quit = mQuit, sleep = mLandTipSchedulingSleep, this]() -> ll::coro::CoroTask<> {
-            while (!quit->load()) {
-                co_await sleep->sleepFor(Config::cfg.land.tip.bottomTipFrequency * ll::chrono::ticks{20});
-                if (quit->load()) {
-                    break;
-                }
-
-                if (mLandIdMap.empty()) {
-                    continue;
-                }
-
-                try {
-                    tickLandTip();
-                } catch (std::exception& e) {
-                    PLand::getInstance().getSelf().getLogger().error(
-                        "An exception occurred while scheduling land tip: {}",
-                        e.what()
-                    );
-                } catch (...) {
-                    PLand::getInstance().getSelf().getLogger().error(
-                        "An unknown exception occurred while scheduling land tip."
-                    );
-                }
-            }
-        }).launch(ll::thread::ServerThreadExecutor::getDefault());
-    }
-}
-
-LandScheduler::~LandScheduler() {
-    auto& bus = ll::event::EventBus::getInstance();
-    bus.removeListener(mPlayerEnterLandListener);
-    bus.removeListener(mPlayerJoinServerListener);
-    bus.removeListener(mPlayerDisconnectListener);
-    mQuit->store(true);
-    mEventSchedulingSleep->interrupt(true);
-    mLandTipSchedulingSleep->interrupt(true);
-    mPlayers.clear();
-    mDimensionMap.clear();
-    mLandIdMap.clear();
-}
-
-void LandScheduler::tickEvent() {
-    auto& bus      = ll::event::EventBus::getInstance();
-    auto& registry = PLand::getInstance().getLandRegistry();
-
-    auto iter = mPlayers.begin();
-    while (iter != mPlayers.end()) {
-        try {
-            auto player = *iter;
-
-            auto const& currentPos   = player->getPosition();
-            int const   currentDimId = player->getDimensionId();
-
-            int&  lastDimId  = this->mDimensionMap[player];
-            auto& lastLandID = this->mLandIdMap[player];
-
-            auto   land          = registry.getLandAt(currentPos, currentDimId);
-            LandID currentLandId = land ? land->getId() : INVALID_LAND_ID;
-
-            // 处理维度变化
-            if (currentDimId != lastDimId) {
-                if (lastLandID != INVALID_LAND_ID) {
-                    bus.publish(event::PlayerLeaveLandEvent{*player, lastLandID}); // 离开上一个维度的领地
-                }
-                lastDimId = currentDimId;
-            }
-
-            // 处理领地变化
-            if (currentLandId != lastLandID) {
-                if (lastLandID != INVALID_LAND_ID) {
-                    bus.publish(event::PlayerLeaveLandEvent{*player, lastLandID}); // 离开上一个领地
-                }
-                if (currentLandId != INVALID_LAND_ID) {
-                    bus.publish(event::PlayerEnterLandEvent{*player, currentLandId}); // 进入新领地
-                }
-                lastLandID = currentLandId;
-            }
-            ++iter;
-        } catch (...) {
-            iter = mPlayers.erase(iter);
-        }
-    }
-}
-
-void LandScheduler::tickLandTip() {
-    auto& playerInfo = ll::service::PlayerInfo::getInstance();
-    auto& registry   = PLand::getInstance().getLandRegistry();
-
-    SetTitlePacket pkt(SetTitlePacket::TitleType::Actionbar);
-    for (auto& [player, landId] : mLandIdMap) {
-        if (landId == INVALID_LAND_ID) {
-            continue;
-        }
-
-        if (auto settings = registry.getPlayerSettings(player->getUuid());
-            settings && !settings->showBottomContinuedTip) {
-            continue; // 如果玩家设置不显示底部提示，则跳过
-        }
-
-        auto land = registry.getLand(landId);
-        if (!land) {
-            continue;
-        }
-
-        auto& owner = land->getOwner();
-        auto  info  = playerInfo.fromUuid(owner);
-        if (land->isOwner(player->getUuid())) {
-            pkt.mTitleText = "[Land] 当前正在领地 {}"_trf(*player, land->getName());
-        } else {
-            pkt.mTitleText = "[Land] 这里是 {} 的领地"_trf(*player, info.has_value() ? info->name : owner.asString());
-        }
-
-        pkt.sendTo(*player);
-    }
-}
-
-
-} // namespace land
